+++ src/main/java/com/nguyenquyhy/discordbridge/DiscordBridge.java
@@ -5,6 +5,7 @@ import com.nguyenquyhy.discordbridge.database.IStorage;
 import com.nguyenquyhy.discordbridge.listeners.ChatListener;
 import com.nguyenquyhy.discordbridge.listeners.ClientConnectionListener;
 import com.nguyenquyhy.discordbridge.listeners.DeathListener;
+import com.nguyenquyhy.discordbridge.listeners.FirstJoinListener;
 import com.nguyenquyhy.discordbridge.logics.ConfigHandler;
 import com.nguyenquyhy.discordbridge.logics.LoginHandler;
 import com.nguyenquyhy.discordbridge.models.ChannelConfig;
@@ -20,6 +21,7 @@ import org.spongepowered.api.Game;
 import org.spongepowered.api.Sponge;
 import org.spongepowered.api.config.ConfigDir;
 import org.spongepowered.api.event.Listener;
+import org.spongepowered.api.event.game.state.GamePostInitializationEvent;
 import org.spongepowered.api.event.game.state.GamePreInitializationEvent;
 import org.spongepowered.api.event.game.state.GameStartedServerEvent;
 import org.spongepowered.api.event.game.state.GameStoppingServerEvent;
@@ -35,7 +37,7 @@ import java.util.*;
 @Plugin(
     id = "discordbridge",
     name = "Discord Bridge",
-    version = "2.4.1",
+    version = "2.4.2",
     description = "A Sponge plugin to connect your Minecraft server with Discord",
     authors = {"Hy", "Mohron"}
 )
@@ -67,10 +69,17 @@ public class DiscordBridge {
     public void onPreInitialization(GamePreInitializationEvent event) throws IOException, ObjectMappingException {
         instance = this;
         config = ConfigHandler.loadConfiguration();
+    }
 
+    @Listener
+    public void onPostInitialization(GamePostInitializationEvent event) {
         Sponge.getEventManager().registerListeners(this, new ChatListener());
         Sponge.getEventManager().registerListeners(this, new ClientConnectionListener());
         Sponge.getEventManager().registerListeners(this, new DeathListener());
+
+        if (Sponge.getPluginManager().isLoaded("nucleus")) {
+            Sponge.getEventManager().registerListeners(this, new FirstJoinListener());
+        }
     }
 
     @Listener